3_legal_translation/requirements.txt
langchain
langchain-openai
langchain-community
python-dotenv
streamlit
tiktoken
pandas
